Move Html component into website middleware
