+++ packages/server/package.json
@@ -196,7 +196,6 @@
     "resolve-url-loader": "^2.2.1",
     "rimraf": "^2.6.2",
     "sass-loader": "^7.0.1",
-    "serialize-javascript": "^1.5.0",
     "source-list-map": "^2.0.0",
     "style-loader": "^0.21.0",
     "url-loader": "^1.0.1",
