--- packages/server/src/middleware/html.tsx
@@ -1,60 +1,0 @@
-import React, { ReactElement } from 'react';
-import { HelmetData } from 'react-helmet';
-import serialize from 'serialize-javascript';
-import clientModules from '../../../client/src/modules';
-import { styles } from '../../../client/src/modules/common/components/web';
-
-interface HtmlProps {
-  content: string;
-  state: any;
-  assetMap: { [key: string]: string };
-  css: Array<ReactElement<{}>>;
-  helmet: HelmetData;
-}
-
-export default ({ content, state, assetMap, css, helmet }: HtmlProps) => (
-  <html lang="en" {...helmet.htmlAttributes.toComponent()}>
-    <head>
-      {helmet.title.toComponent()}
-      {helmet.meta.toComponent()}
-      {helmet.link.toComponent()}
-      <meta charSet="utf-8" />
-      <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
-      <link rel="apple-touch-icon" sizes="180x180" href={`${assetMap['apple-touch-icon.png']}`} />
-      <link rel="icon" type="image/png" href={`${assetMap['favicon-32x32.png']}`} sizes="32x32" />
-      <link rel="icon" type="image/png" href={`${assetMap['favicon-16x16.png']}`} sizes="16x16" />
-      <link rel="manifest" href={`${assetMap['manifest.xjson']}`} />
-      <link rel="mask-icon" href={`${assetMap['safari-pinned-tab.svg']}`} color="#5bbad5" />
-      <link rel="shortcut icon" href={`${assetMap['favicon.ico']}`} />
-      <meta name="msapplication-config" content={`${assetMap['browserconfig.xml']}`} />
-      <meta name="theme-color" content="#ffffff" />
-      {!__DEV__ && <link rel="stylesheet" type="text/css" href={`${assetMap['index.css']}`} />}
-      {!!__DEV__ && (
-        <style
-          dangerouslySetInnerHTML={{
-            __html: styles._getCss() + clientModules.stylesInserts.map(style => style._getCss()).join('')
-          }}
-        />
-      )}
-      {!!css && css}
-      {clientModules.scriptsInserts.map((script, i) => {
-        if (script) {
-          return <script key={i} src={script} />;
-        }
-      })}
-    </head>
-    <body {...helmet.bodyAttributes.toComponent()}>
-      <div id="root" dangerouslySetInnerHTML={{ __html: content || '' }} />
-      <script
-        dangerouslySetInnerHTML={{
-          __html: `window.__APOLLO_STATE__=${serialize(state, {
-            isJSON: true
-          })};`
-        }}
-        charSet="UTF-8"
-      />
-      {assetMap['vendor.js'] && <script src={`${assetMap['vendor.js']}`} charSet="utf-8" />}
-      <script src={`${assetMap['index.js']}`} charSet="utf-8" />
-    </body>
-  </html>
-);
