+++ packages/server/src/middleware/website.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactElement } from 'react';
 import ReactDOMServer from 'react-dom/server';
 import { SchemaLink } from 'apollo-link-schema';
 import { ApolloProvider, getDataFromTree } from 'react-apollo';
@@ -7,17 +7,73 @@ import { StaticRouter } from 'react-router';
 import { ServerStyleSheet } from 'styled-components';
 import fs from 'fs';
 import path from 'path';
-import Helmet from 'react-helmet';
+import Helmet, { HelmetData } from 'react-helmet';
+import serialize from 'serialize-javascript';
 
 import { isApiExternal, apiUrl } from '../net';
 import createApolloClient from '../../../common/createApolloClient';
 import createReduxStore from '../../../common/createReduxStore';
-import Html from './html';
 import Routes from '../../../client/src/app/Routes';
 import modules from '../modules';
 import schema from '../api/schema';
+import { styles } from '../../../client/src/modules/common/components/web';
 
 let assetMap: { [key: string]: string };
+
+interface HtmlProps {
+  content: string;
+  state: any;
+  clientModules: any;
+  css: Array<ReactElement<{}>>;
+  helmet: HelmetData;
+}
+
+const Html = ({ content, state, css, clientModules, helmet }: HtmlProps) => (
+  <html lang="en" {...helmet.htmlAttributes.toComponent()}>
+    <head>
+      {helmet.title.toComponent()}
+      {helmet.meta.toComponent()}
+      {helmet.link.toComponent()}
+      <meta charSet="utf-8" />
+      <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
+      <link rel="apple-touch-icon" sizes="180x180" href={`${assetMap['apple-touch-icon.png']}`} />
+      <link rel="icon" type="image/png" href={`${assetMap['favicon-32x32.png']}`} sizes="32x32" />
+      <link rel="icon" type="image/png" href={`${assetMap['favicon-16x16.png']}`} sizes="16x16" />
+      <link rel="manifest" href={`${assetMap['manifest.xjson']}`} />
+      <link rel="mask-icon" href={`${assetMap['safari-pinned-tab.svg']}`} color="#5bbad5" />
+      <link rel="shortcut icon" href={`${assetMap['favicon.ico']}`} />
+      <meta name="msapplication-config" content={`${assetMap['browserconfig.xml']}`} />
+      <meta name="theme-color" content="#ffffff" />
+      {!__DEV__ && <link rel="stylesheet" type="text/css" href={`${assetMap['index.css']}`} />}
+      {!!__DEV__ && (
+        <style
+          dangerouslySetInnerHTML={{
+            __html: styles._getCss() + clientModules.stylesInserts.map((style: any) => style._getCss()).join('')
+          }}
+        />
+      )}
+      {!!css && css}
+      {clientModules.scriptsInserts.map((script: string, i: number) => {
+        if (script) {
+          return <script key={i} src={script} />;
+        }
+      })}
+    </head>
+    <body {...helmet.bodyAttributes.toComponent()}>
+      <div id="root" dangerouslySetInnerHTML={{ __html: content || '' }} />
+      <script
+        dangerouslySetInnerHTML={{
+          __html: `window.__APOLLO_STATE__=${serialize(state, {
+            isJSON: true
+          })};`
+        }}
+        charSet="UTF-8"
+      />
+      {assetMap['vendor.js'] && <script src={`${assetMap['vendor.js']}`} charSet="utf-8" />}
+      <script src={`${assetMap['index.js']}`} charSet="utf-8" />
+    </body>
+  </html>
+);
 
 const renderServerSide = async (req: any, res: any) => {
   const clientModules = require('../../../client/src/modules').default;
@@ -57,12 +113,12 @@ const renderServerSide = async (req: any, res: any) => {
     }
 
     const sheet = new ServerStyleSheet();
-    const htmlProps = {
+    const htmlProps: HtmlProps = {
       content: ReactDOMServer.renderToString(sheet.collectStyles(App)),
       css: sheet.getStyleElement().map((el, idx) => (el ? React.cloneElement(el, { key: idx }) : el)),
       helmet: Helmet.renderStatic(), // Avoid memory leak while tracking mounted instances
       state: { ...client.cache.extract() },
-      assetMap
+      clientModules
     };
 
     res.send(`<!doctype html>\n${ReactDOMServer.renderToStaticMarkup(<Html {...htmlProps} />)}`);
